+++ source/kernel/src/drivers/mod.rs
@@ -1,2 +1,3 @@
+pub mod emb_display;
 pub mod serial_mux;
 pub mod simple_serial;
